sched/isolation: add 'inverse' parameter for the 'nohz_full' and 'isolcpus'
By default 'nohz_full' and 'isolcpus' accept a range of those CPUs,
which should be isolated from the housekeeping, with 'inverse' the
logic is inverted: you specify CPUs for housekeeping tasks, others
are isolated.

The motivation of the patch is to generalize and simplify kernel
configration, when maximum number of CPUs is not known, but it is
known, that CPU0 will do the whole housekeeping, so the command
line becomes:

    isolcpus=inverse,0 nohz_full=inverse,0

instead of

   isolcpus=1-N nohz_full=1-N

where N is the index of the maximum CPU, which is not always known.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>