configs: Disable INTEL_PCH_THERMAL driver
The INTEL_PCH_THERMAL driver was enabled previously along with other thermal
device drivers to improve thermal monitoring capabilities. However, on some
devices this particular driver will make the thermal device in the PCIe bus to
be enumerated in the same IOMMU group of USB Controller regardless the ACS
patch, which break some use cases where USB Controller is passed-through to
VMs.

Let's keep this driver disabled for now (as it was before), until we find a
better way to mitigate the issue for the affected devices.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>